requirements.txt
transformers==4.37.2
torch==2.1.2
diffusers==0.25.0
accelerate==0.26.1
safetensors==0.4.2
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
pillow==10.2.0
pydantic==2.6.1 